refactor: 필터링 관련 코드 구조 변경
- 우선은 구조만 바꿨습니다. (가까운 매칭 찾기와 동일한 구성)
- 추후 리팩토링할 예정입니다.
